models/requirements.txt
annotated-types==0.5.0
argcomplete==1.10.3
asttokens==2.4.0
backcall==0.2.0
beautifulsoup4==4.8.2
blis==0.7.11
catalogue==2.0.9
certifi==2023.7.22
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.2.0
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
compressed-rtf==1.0.6
confection==0.1.3
configobj==5.0.8
configparser==6.0.0
cryptography==41.0.4
cymem==2.0.8
decorator==5.1.1
defusedxml==0.7.1
docx2txt==0.8
ebcdic==1.1.1
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.6.0/en_core_web_md-3.6.0-py3-none-any.whl#sha256=1618fc62e7aedc2b778caa3ea6e9c8280beac3b83fe7b93316cbb9e8f3abe7c5
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl#sha256=83276fc78a70045627144786b52e1f2728ad5e29e5e43916ec37ea9c26a11212
etelemetry==0.3.0
exceptiongroup==1.1.3
executing==1.2.0
extract-msg==0.28.7
ezodf==0.3.2
filelock==3.12.4
fitz==0.0.1.dev2
future==0.18.3
httplib2==0.22.0
idna==3.4
IMAPClient==2.1.0
ipython==8.15.0
isodate==0.6.1
jedi==0.19.0
jellyfish==1.0.1
Jinja2==3.1.2
joblib==1.3.2
keras==2.14.0
langcodes==3.3.0
looseversion==1.3.0
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
murmurhash==1.0.10
networkx==3.1
nibabel==5.1.0
nipype==1.8.6
nltk==3.8.1
numpy==1.26.0
odfpy==1.4.1
olefile==0.46
packaging==23.1
pandas==2.1.1
parso==0.8.3
pathlib==1.0.1
pathy==0.10.2
pdfminer.six==20191110
pickleshare==0.7.5
Pillow==9.5.0
preshed==3.0.9
prompt-toolkit==3.0.39
prov==2.0.0
pure-eval==0.2.2
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.3.0
pydantic_core==2.6.3
pydot==1.4.2
Pygments==2.16.1
pyparsing==3.1.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-docx==0.8.11
python-pptx==0.6.22
pytz==2023.3.post1
pyxnat==1.6
rdflib==7.0.0
regex==2023.8.8
requests==2.31.0
scipy==1.11.3
simplejson==3.19.2
six==1.16.0
skillNer==1.0.3
smart-open==6.4.0
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.8.1
srsly==2.4.8
stack-data==0.6.2
textract==1.6.5
thinc==8.1.12
tqdm==4.66.1
traitlets==5.10.1
traits==6.3.2
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.5
wasabi==1.1.2
wcwidth==0.2.6
xlrd==1.2.0
XlsxWriter==3.1.5
